change how to handle device specific config
before I'd have device specific configs versioned in the repo. I think
this is fine for me, but is confusing for others coming to this repo as
a generic way to install arch and configure hyprland, which is how this
repo advertises itself. I'm falling back to preserving the config files
for each of my devices with my backup strategy