add custom hyprshade presets for yoga laptop
bluelight filter and vibrance are respectively too aggressive and too
modest for the lenovo yoga slim 7i display. These custom shaders set
temp to 4000 and tone down reds a bit, so that the colors don't look
too oversaturated. It'd be nice to tone reds even further, but it starts
affecting the balance of other colors, so this is a nice middleground